Refactored MssSystemTablesTest
